Use RGB color ordering
The LEDs never were taking the values in GRB order. However, the library
thought they were (because of a tricky default), so it was changing the
order before sending it to the LEDs, which made it look like you had to
set the colors in GRB order. You never have to if you initialize the
library with the correct color ordering.

This initializes the library with the right color order setting, and
removes the work-around.