Updated run-folder.py
Rewritten to parse the CSV correctly and support variable FRAME_TIME
The lerping code had to be rewritten because the old way did not support variable frame times. The new way just interpolates between the last and first frames if they are different enough.
Moved the CLI inputs to argparse